requirements.txt
flask
qrcode
Pillow
cryptography
pyotp